requirements.txt
-f /usr/share/pip-wheels
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cycler==0.10.0
Flask==1.1.2
h5py==2.10.0
idna==2.9
imageai==2.1.5
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==7.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
PyYAML==5.3.1
scipy==1.4.1
six==1.14.0
soupsieve==2.0
Werkzeug==1.0.1
